LPD-58182 Inline GeoShapeQueryTranslator/GeoShapeQueryTranslatorImpl
